Refactor & Clean Up to plugins/odp
